Extract finally-dispose compiler fixtures
